✨ Don't comment on dependabot reviews
It turns out that the GITHUB_TOKEN that is used for dependabot actions
has no write actions at all, which includes making comments on a PR.
As a result, we can't comment at the end of a workflow that's started by
dependabot (e.g. PRs), so we just do nothing, and rely on the exit code
being zero (i.e. the workflow succeeding) to allow for merges.