HPC-8345: Remove usages of `String.prototype.substr()`
String.prototype.substr() was never part of
the main ECMAScript specification and needs
to be replaced by a ECMA-standard
String.prototype.substring().

The two parameters of `substr()` are
`start` and `length`, while for `substring()`,
they are `start` and `end`.

This means that when there is a single
argument being used, both methods are
behaving the same. With two arguments,
they behave different, but there is only
one single usage and first argument is
zero, so both methods boil down to same
set of arguments essentially.