Allow for multiple docker images to be built
Allow to specify multiple docker images to be built,
essentially allowing for a monorepo approach.
Since there is a change in configuration, some config
options are renamed to force clients to update.

The strategy is to use a single git tag when a
new deployment happens and all applications inside
a monorepo should share the same version. This
version will be used to name a single git tag
and to tag multiple Docker images being built
and released.

In cases when multiple applications are deployed,
multiple deployments will be created on GitHub,
one for each app.