Accept "apps to build" as action input param
In order to avoid building and releasing
unchanged apps in monorepo setups, the
action now accepts input parameter, specifying
which apps to build and release.
Input parameter specified in `action.yml`
is available through env variable.

In order to know which options are valid,
we also allow configuration of buildable
Docker images to specify name of the app
for which they are configured.