Set app name in deployment payload
Jenkins job which is triggered via GitHub webhook
when a deployment is created on GitHub reads from
this payload to know which image tag to deploy -
via `docker_tag` parameter.

This same job assumes one deployable piece of
software per repo, thus reads repository name
to know which image to deploy. In monorepos,
there are multiple deployable pieces of software,
thus we need a way to differentiate what needs
to be deployed. Hence, this new payload variable.