Send repository name in deployment payload
First, strongly type the deployment payload
object and don't allow just any key.
`docker_tag` was the only used payload so
far and we add `repository_name` now.
This repository name payload is needed when
deploying from a monorepo, for ticket HPC-8676.

Jenkins job that gets triggered by deployment
expects repository name to match application
that is being deployed. Now, to allow releasing
multiple applications from the same repo, we
need to specify the application name in payload.

But, since deployment from a monorepo is only
needed for applications and not all repositories
are monorepo ones, we need to keep old behavior,
that is why this payload defaults to repo name.