Be specific in ignorefiles pattern format
According to official docs:
https://git-scm.com/docs/gitignore#_pattern_format
> If there is a separator at the beginning or
> middle (or both) of the pattern, then the pattern
> is relative to the directory level of the particular
> `.gitignore` file itself. Otherwise the pattern may
> also match at any level below the `.gitignore` level.

Because there is no separator at the beggining of
`dist/` in `.prettierignore`, it would match other
directories named `dist` elsewhere, not only the
one existing on the same level as `.prettierignore`.

That could be what we want, but I want to be specific
about what we exclude and later add more things if
needed and still be specific about them.